migrate a couple tests
