+++ apps/admin/frontend/src/screens/write_ins_summary_screen.test.tsx
@@ -1,13 +1,9 @@
 import userEvent from '@testing-library/user-event';
 import { electionMinimalExhaustiveSampleFixtures } from '@votingworks/fixtures';
 import { sleep } from '@votingworks/basics';
-import type {
-  WriteInCandidateRecord,
-  WriteInRecordPending,
-  WriteInPendingTally,
-} from '@votingworks/admin-backend';
+import type { WriteInPendingTally } from '@votingworks/admin-backend';
 import { createMemoryHistory } from 'history';
-import { act, screen, waitFor } from '../../test/react_testing_library';
+import { act, screen } from '../../test/react_testing_library';
 import { renderInAppContext } from '../../test/render_in_app_context';
 import { WriteInsSummaryScreen } from './write_ins_summary_screen';
 import { ApiMock, createApiMock } from '../../test/helpers/api_mock';
@@ -15,16 +11,6 @@ import { ApiMock, createApiMock } from '../../test/helpers/api_mock';
 jest.setTimeout(20000);
 
 const { electionDefinition } = electionMinimalExhaustiveSampleFixtures;
-
-function mockWriteInRecordPending(id: string): WriteInRecordPending {
-  return {
-    id,
-    contestId: 'zoo-council-mammal',
-    optionId: 'write-in-0',
-    castVoteRecordId: 'id',
-    status: 'pending',
-  };
-}
 
 function mockWriteInTallyPending(
   contestId: string,
@@ -106,223 +92,4 @@ test('CVRs with write-ins loaded', async () => {
   expect(history.location.pathname).toEqual(
     '/write-ins/adjudication/zoo-council-mammal'
   );
-});
-
-test('ballot pagination', async () => {
-  const contestId = 'zoo-council-mammal';
-  const pageCount = 3;
-  apiMock.expectGetWriteInTallies([mockWriteInTallyPending(contestId, 3)]);
-  apiMock.expectGetCastVoteRecordFiles([]);
-
-  renderInAppContext(<WriteInsSummaryScreen />, {
-    electionDefinition,
-    apiMock,
-  });
-
-  const mockWriteInRecords = ['0', '1', '2'].map(mockWriteInRecordPending);
-  apiMock.expectGetWriteIns(mockWriteInRecords, contestId);
-  apiMock.expectGetWriteInCandidates([], contestId);
-  for (const mockWriteInRecord of mockWriteInRecords) {
-    apiMock.expectGetWriteInDetailView(mockWriteInRecord.id);
-  }
-
-  userEvent.click(await screen.findByText(`Adjudicate ${pageCount}`));
-
-  for (let pageNumber = 1; pageNumber <= pageCount; pageNumber += 1) {
-    await screen.findByText(new RegExp(`${pageNumber} of ${pageCount}`));
-    const previousButton = await screen.findButton('Previous');
-    if (pageNumber === 1) {
-      expect(previousButton).toBeDisabled();
-    } else {
-      expect(previousButton).not.toBeDisabled();
-    }
-
-    const nextButton = await screen.findButton('Next');
-    if (pageNumber === pageCount) {
-      expect(nextButton).toBeDisabled();
-      const doneButton = await screen.findButton('Back to All Write-Ins');
-      doneButton.click();
-    } else {
-      expect(nextButton).not.toBeDisabled();
-      nextButton.click();
-    }
-  }
-});
-
-test('adjudication', async () => {
-  const contestId = 'zoo-council-mammal';
-  apiMock.expectGetWriteInTallies([mockWriteInTallyPending(contestId, 2)]);
-  apiMock.expectGetCastVoteRecordFiles([]);
-
-  renderInAppContext(<WriteInsSummaryScreen />, {
-    electionDefinition,
-    apiMock,
-  });
-
-  // open adjudication screen
-  const mockWriteInRecords = ['0', '1'].map(mockWriteInRecordPending);
-  apiMock.expectGetWriteIns(mockWriteInRecords, contestId);
-  const mockWriteInCandidate: WriteInCandidateRecord = {
-    id: 'lemur',
-    name: 'Lemur',
-    contestId: 'zoo-council-mammal',
-    electionId: 'id',
-  };
-  apiMock.expectGetWriteInCandidates([mockWriteInCandidate], contestId);
-  apiMock.expectGetWriteInDetailView(mockWriteInRecords[0].id);
-  apiMock.expectGetWriteInDetailView(mockWriteInRecords[1].id); // prefetch
-  userEvent.click(await screen.findByText('Adjudicate 2'));
-
-  await screen.findByRole('img', { name: /ballot/i });
-  screen.getButton('Zebra');
-  screen.getButton('Lion');
-  screen.getButton('Kangaroo');
-  screen.getButton('Elephant');
-  screen.getButton('Lemur');
-
-  // adjudicate for official candidate
-  const partialMockAdjudicatedWriteIn = {
-    id: '0',
-    contestId: 'zoo-council-mammal',
-    optionId: 'write-in-0',
-    castVoteRecordId: 'id',
-    status: 'adjudicated',
-  } as const;
-  apiMock.apiClient.adjudicateWriteIn
-    .expectCallWith({
-      writeInId: mockWriteInRecords[0].id,
-      type: 'official-candidate',
-      candidateId: 'zebra',
-    })
-    .resolves();
-  apiMock.expectGetWriteIns(
-    [
-      {
-        ...partialMockAdjudicatedWriteIn,
-        adjudicationType: 'official-candidate',
-        candidateId: 'zebra',
-      },
-      mockWriteInRecords[1],
-    ],
-    contestId
-  );
-  apiMock.expectGetWriteInCandidates([mockWriteInCandidate], contestId);
-  apiMock.expectGetWriteInDetailView(mockWriteInRecords[1].id);
-  apiMock.expectGetWriteInTallies([]);
-
-  userEvent.click(screen.getButton('Zebra'));
-  await waitFor(async () =>
-    expect(await screen.findButton('Next')).toHaveFocus()
-  );
-
-  // clicking current selection should be no-op
-  userEvent.click(screen.getButton('Zebra'));
-  apiMock.assertComplete();
-
-  // adjudicate for existing write-in candidate
-  apiMock.apiClient.adjudicateWriteIn
-    .expectCallWith({
-      writeInId: mockWriteInRecords[0].id,
-      type: 'write-in-candidate',
-      candidateId: 'lemur',
-    })
-    .resolves();
-  apiMock.expectGetWriteIns(
-    [
-      {
-        ...partialMockAdjudicatedWriteIn,
-        adjudicationType: 'write-in-candidate',
-        candidateId: 'lemur',
-      },
-      mockWriteInRecords[1],
-    ],
-    contestId
-  );
-  apiMock.expectGetWriteInCandidates([mockWriteInCandidate], contestId);
-  apiMock.expectGetWriteInDetailView(mockWriteInRecords[1].id);
-  apiMock.expectGetWriteInTallies([]);
-  userEvent.click(screen.getButton('Lemur'));
-  await waitFor(async () =>
-    expect(await screen.findButton('Next')).toHaveFocus()
-  );
-
-  // clicking current selection should be no-op
-  userEvent.click(screen.getButton('Lemur'));
-  apiMock.assertComplete();
-
-  // adjudicate for a new write-in-candidate
-  const mockNewWriteInCandidateRecord: WriteInCandidateRecord = {
-    id: 'dh',
-    contestId,
-    electionId: 'any',
-    name: 'Dark Helmet',
-  };
-  apiMock.apiClient.addWriteInCandidate
-    .expectCallWith({ contestId, name: 'Dark Helmet' })
-    .resolves(mockNewWriteInCandidateRecord);
-  apiMock.apiClient.adjudicateWriteIn
-    .expectCallWith({
-      writeInId: mockWriteInRecords[0].id,
-      type: 'write-in-candidate',
-      candidateId: 'dh',
-    })
-    .resolves();
-  apiMock.expectGetWriteIns(
-    [
-      {
-        ...partialMockAdjudicatedWriteIn,
-        adjudicationType: 'official-candidate',
-        candidateId: 'dh',
-      },
-      mockWriteInRecords[1],
-    ],
-    contestId
-  );
-  apiMock.expectGetWriteInCandidates(
-    [mockWriteInCandidate, mockNewWriteInCandidateRecord],
-    contestId
-  );
-  apiMock.expectGetWriteInCandidates(
-    [mockWriteInCandidate, mockNewWriteInCandidateRecord],
-    contestId
-  );
-  apiMock.expectGetWriteInDetailView(mockWriteInRecords[1].id);
-  apiMock.expectGetWriteInTallies([]);
-  userEvent.click(await screen.findButton(/Add New Write-In Candidate/i));
-  userEvent.type(
-    await screen.findByPlaceholderText('Candidate Name'),
-    'Dark Helmet'
-  );
-  userEvent.click(await screen.findByText('Add'));
-
-  await screen.findButton('Dark Helmet');
-  await waitFor(async () =>
-    expect(await screen.findButton('Next')).toHaveFocus()
-  );
-
-  // adjudicate as invalid
-  apiMock.apiClient.adjudicateWriteIn
-    .expectCallWith({
-      writeInId: mockWriteInRecords[0].id,
-      type: 'invalid',
-    })
-    .resolves();
-  apiMock.expectGetWriteIns(
-    [
-      {
-        ...partialMockAdjudicatedWriteIn,
-        adjudicationType: 'invalid',
-      },
-      mockWriteInRecords[1],
-    ],
-    contestId
-  );
-  apiMock.expectGetWriteInCandidates([mockWriteInCandidate], contestId);
-  apiMock.expectGetWriteInDetailView(mockWriteInRecords[1].id);
-  apiMock.expectGetWriteInTallies([]);
-  userEvent.click(await screen.findButton(/Mark Write-In Invalid/i));
-  await waitFor(async () =>
-    expect(await screen.findButton('Next')).toHaveFocus()
-  );
-  expect(screen.queryByText('Dark Helmet')).not.toBeInTheDocument();
 });
